Add files via upload
Current release of War GUI version, which is the updated version I made of the War CONSOLE version. 

This War Card Game project (GUI and CONSOLE) is an old project. 
There are some updates to the War instance handling, and some cleanliness changes I would make, but this is the current status for now.